# api/pkg/routes/server.gen.go
// Package routes provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.15.0 DO NOT EDIT.
package routes

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

// Metadata defines model for Metadata.
type Metadata struct {
	Description string `json:"description"`
	Image       string `json:"image"`
	Title       string `json:"title"`
	Url         string `json:"url"`
}

// GetMetadataParams defines parameters for GetMetadata.
type GetMetadataParams struct {
	// Url The URL for which you want to retrieve OpenGraph data.
	Url string `form:"url" json:"url"`
}

// OpenGraphParams defines parameters for OpenGraph.
type OpenGraphParams struct {
	// Url The URL for which you want to retrieve OpenGraph data.
	Url string `form:"url" json:"url"`

	// Title Optional custom title to use for OpenGraph data.
	Title *string `form:"title,omitempty" json:"title,omitempty"`

	// Description Optional custom description to use for OpenGraph data.
	Description *string `form:"description,omitempty" json:"description,omitempty"`

	// Image Optional custom image to use for OpenGraph data.
	Image *string `form:"image,omitempty" json:"image,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get metadata of a URL
	// (GET /metadata)
	GetMetadata(ctx echo.Context, params GetMetadataParams) error
	// OpenGraph Data
	// (GET /opengraph)
	OpenGraph(ctx echo.Context, params OpenGraphParams) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetMetadata converts echo context to params.
func (w *ServerInterfaceWrapper) GetMetadata(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetMetadataParams
	// ------------- Required query parameter "url" -------------

	err = runtime.BindQueryParameter("form", true, true, "url", ctx.QueryParams(), &params.Url)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter url: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetMetadata(ctx, params)
	return err
}

// OpenGraph converts echo context to params.
func (w *ServerInterfaceWrapper) OpenGraph(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params OpenGraphParams
	// ------------- Required query parameter "url" -------------

	err = runtime.BindQueryParameter("form", true, true, "url", ctx.QueryParams(), &params.Url)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter url: %s", err))
	}

	// ------------- Optional query parameter "title" -------------

	err = runtime.BindQueryParameter("form", true, false, "title", ctx.QueryParams(), &params.Title)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter title: %s", err))
	}

	// ------------- Optional query parameter "description" -------------

	err = runtime.BindQueryParameter("form", true, false, "description", ctx.QueryParams(), &params.Description)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter description: %s", err))
	}

	// ------------- Optional query parameter "image" -------------

	err = runtime.BindQueryParameter("form", true, false, "image", ctx.QueryParams(), &params.Image)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter image: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.OpenGraph(ctx, params)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/metadata", wrapper.GetMetadata)
	router.GET(baseURL+"/opengraph", wrapper.OpenGraph)

}